Include lunr.min.js in this repository
This is necessary if we want to be able to install this gem from GitHub,
without first creating a package and sticking it on rubygems.org. The
package creation code copies lunr.min.js to this location, so it gets
bundled into the gem. But if you install directly from the source code,
that package never gets built, and so that step is missing. Downstream
then breaks because the lunr.min.js file isn't where it is expected to
be.

This is definitely a hack, but trying for now to get things working.